requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
jinja2==3.1.3
python-dotenv==1.0.1
aiofiles==23.2.1
